New block size algorithm
This adjusts transaction sizes that can be submitted and pulled into a block template from the transaction pool. Some transaction could be accepted by the daemon/wallet being far to huge for the current height of the blockchain. Therefore the algorithm to determine the maximum size of transactions that will be committed to the transaction pool has been adjusted. 
In some circumstances the predicted median size yielded invalid block templates due to too large transactions embedded. Therefore the algorithm choosing the transaction to include into a new block has been adjusted to also verify that the maximum block size wont be exceeded once the mined block gets committed.